Added dry-run mode
When testing the building of indexes, it is desirable to stop Inquisito
from making the HTTP POST into Amazon CloudSearch. This commit introduces
a 'dry run' configuration parameter that prevents these calls from being
made. Otherwise, Inquisito behaves as normal and will log out all of the SDF
that it would have sent to CloudSearch.